refactor : Map Configuration 분리
